requirements.txt
# QuantBot Dependencies
ccxt>=4.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
requests>=2.31.0
flask>=3.0.0